Moved PlatformLock to Internals namespace + added ability to plugin custom factory
